+++ changelog.md
@@ -1,5 +1,12 @@
 # 更新日志
 
+## 1.23.2
+
+- 【修复】切换环境请求头不会变化
+- 【修复】分享出来的Markdown文档不显示 [#I7SJ3O](https://gitee.com/durcframework/torna/issues/I7SJ3O)
+- 【修复】点击锁定文档按钮，锁图标没显示 [#I7SJ3X](https://gitee.com/durcframework/torna/issues/I7SJ3X)
+- 【修复】导入swagger basic认证失败 [#I7S3LU](https://gitee.com/durcframework/torna/issues/I7S3LU)
+
 ## 1.23.1
 
 - 【修复】内容覆盖更新配置不生效问题